Moved shapes to geometry
