+++ src/frontend/geometry/curve.cpp
