+++ src/frontend/geometry/curve.h
