+++ src/frontend/geometry/geometry.cpp
