+++ src/frontend/geometry/trianglemesh.cpp
